Secure SSL defaults, split external.py for maintainability
- Change verify_ssl default to True (secure by default)
- Extract output generation to external_output.py (607 lines)
- Reduce external.py from 1720 to 1134 lines
- Update tests for new function signatures
- Bump version to 1.6.2
